modularize, document, and minify
